add check if entity use file trait

+++ src/Domain/AbstractAction.php
@@ -11,6 +11,7 @@ use App\Domain\Exceptions\HttpNotFoundException;
 use App\Domain\Exceptions\HttpNotImplementedException;
 use App\Domain\Service\File\FileRelationService;
 use App\Domain\Service\File\FileService;
+use App\Domain\Traits\FileTrait;
 use Illuminate\Support\Collection;
 use Psr\Container\ContainerInterface;
 use Slim\Http\Request;
@@ -182,53 +183,55 @@ abstract class AbstractAction extends AbstractComponent
      */
     protected function processEntityFiles(AbstractEntity $entity, string $field = 'files'): AbstractEntity
     {
-        $fileRelationService = FileRelationService::getWithContainer($this->container);
+        if (in_array(FileTrait::class, class_uses($entity), true)) {
+            $fileRelationService = FileRelationService::getWithContainer($this->container);
 
             // new
             if (($uploaded = $this->getUploadedFiles($field)) !== []) {
                 $index = $entity->hasFiles();
 
-            foreach ($uploaded as $name => $files) {
-                if (is_numeric($name)) {
-                    $name = '';
-                }
+                foreach ($uploaded as $name => $files) {
+                    if (is_numeric($name)) {
+                        $name = '';
+                    }
 
-                foreach ($files as $file) {
-                    $fileRelation = $fileRelationService->create([
-                        'entity' => $entity,
-                        'file' => $file,
-                        'comment' => $name,
-                        'order' => ++$index,
-                    ]);
+                    foreach ($files as $file) {
+                        $fileRelation = $fileRelationService->create([
+                            'entity' => $entity,
+                            'file' => $file,
+                            'comment' => $name,
+                            'order' => ++$index,
+                        ]);
 
-                    // link file to entity
-                    if ($fileRelation) {
-                        $entity->addFile($fileRelation);
+                        // link file to entity
+                        if ($fileRelation) {
+                            $entity->addFile($fileRelation);
+                        }
                     }
                 }
             }
-        }
 
-        // update
-        if (($files = $this->request->getParam($field)) !== null && is_array($files)) {
-            foreach ($files as $uuid => $data) {
-                $default = [
-                    'order' => null,
-                    'comment' => null,
-                    'delete' => null,
-                ];
-                $data = array_merge($default, $data);
+            // update
+            if (($files = $this->request->getParam($field)) !== null && is_array($files)) {
+                foreach ($files as $uuid => $data) {
+                    $default = [
+                        'order' => null,
+                        'comment' => null,
+                        'delete' => null,
+                    ];
+                    $data = array_merge($default, $data);
 
-                $relation = $entity->getFiles()->firstWhere('uuid', $uuid);
+                    $relation = $entity->getFiles()->firstWhere('uuid', $uuid);
 
-                if ($relation) {
-                    if ($data['delete'] !== null) {
-                        $fileRelationService->delete($relation);
+                    if ($relation) {
+                        if ($data['delete'] !== null) {
+                            $fileRelationService->delete($relation);
 
-                        continue;
+                            continue;
+                        }
+
+                        $fileRelationService->update($relation, $data);
                     }
-
-                    $fileRelationService->update($relation, $data);
                 }
             }
         }
@@ -361,7 +364,7 @@ abstract class AbstractAction extends AbstractComponent
      * @throws HttpBadRequestException
      * @throws \RunTracy\Helpers\Profiler\Exception\ProfilerException
      */
-    protected function render($template, array $data = []): string
+    protected function render(string $template, array $data = []): string
     {
         try {
             \RunTracy\Helpers\Profiler\Profiler::start('render (%s)', $template);
